RaspberryPi/requirements.txt
# Requirements for Golf Cart Raspberry Pi Control System

# Core libraries
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0

# GUI libraries
# tkinter is built-in for Python 3, but may need: sudo apt-get install python3-tk

# Serial communication
pyserial>=3.5

# JSON processing (built-in, but explicit version)
# json - built-in

# Math and filtering
scikit-learn>=1.0.0
filterpy>=1.4.5  # For Kalman filtering

# Data handling
pandas>=1.3.0

# Logging and configuration
pyyaml>=6.0
colorlog>=6.0

# Optional: Advanced features
# opencv-python>=4.5.0  # For computer vision if needed later
# flask>=2.0.0          # For web interface if needed
# websockets>=10.0      # For real-time web interface

# Development and testing
pytest>=6.0.0
pytest-mock>=3.6.0

# System monitoring
psutil>=5.8.0

# Mathematical optimization (for path planning)
scipy>=1.7.0
